rename package to adhered to naming convention of other packages in directory
