+++ lib/gear/spatialhandler.py
